updated weekly benefits max (#331)
* updated part-time docs

updated part-time documentation with new maximum weekly amount. Note. Once we push these to a stage env, i need to take screenshots and update the screenshot here

* update UI benefits documentation

updated UI benefits documentation. Once we have the changes in a staging env. i need to take an updated screenshot for Image 5 and update this documentation with that.

* updated weeklyBenefitMax

and the text. Just pinged Katie to confirm that the new change came into effect in "October 2019". If she says otherwise I'll come back and edit this.

* updated variables hyperlink

it was throwing a not found error, so i updated the path

* updated max input

updated the max input amount allowed

* updated max benefits amount

updated max benefits amount

* updated earning disregard formula

with updated max weekly benefits amount

* pull weeklybenefitsmax into a var for ui-parttime

* update screenshot in doc

* update mainnav